Move the frame pacer next to the throttle it works with
The browserd daemon needs the same one-slot keep-latest pacing the
WebMCP frame socket has, and it cannot import a Hono route: it is
bundled into an artifact that runs on a box with nothing but its own
bytes. The alternative was a second copy of logic subtle enough that its
own comments explain why a `bufferedAmount` threshold wedges — exactly
the kind that drifts.

So `createFramePacer` moves verbatim to sit beside `frame-throttle.ts`,
which the daemon already imports for the same reason. No behaviour
change: the route's own 32 tests pass untouched, which is the point of
moving it rather than rewriting it. `toCallbackSocket` stays behind,
being WebSocket-specific.

Deliberately a plain move, with the stall watchdog the daemon needs left
out of it. A watchdog belongs where it can `res.destroy()`, which the
pacer knows nothing about, and adding it here would have changed
behaviour for a shipped socket that does not need it.
